+++ src/edit/electronicAccessFields.js
@@ -16,6 +16,8 @@ import {
   Label,
 } from '@folio/stripes/components';
 
+import { parseEmptyFormValue } from '../utils';
+
 const ElectronicAccessFields = props => {
   const { formatMessage } = useIntl();
 
@@ -73,6 +75,7 @@ const ElectronicAccessFields = props => {
           aria-label={urlRelationshipLabel}
           name={`${field}.relationshipId`}
           component={Select}
+          parse={parseEmptyFormValue}
           dataOptions={[{ label: formatMessage({ id: 'ui-inventory.selectType' }), value: '' }, ...relationshipOptions]}
           disabled={!canEdit}
         />
@@ -82,6 +85,7 @@ const ElectronicAccessFields = props => {
           aria-label={uriLabel}
           name={`${field}.uri`}
           component={TextArea}
+          parse={parseEmptyFormValue}
           rows={1}
           disabled={!canEdit}
         />
@@ -91,6 +95,7 @@ const ElectronicAccessFields = props => {
           aria-label={linkTextLabel}
           name={`${field}.linkText`}
           component={TextArea}
+          parse={parseEmptyFormValue}
           rows={1}
           disabled={!canEdit}
         />
@@ -100,6 +105,7 @@ const ElectronicAccessFields = props => {
           aria-label={materialsSpecificationLabel}
           name={`${field}.materialsSpecification`}
           component={TextArea}
+          parse={parseEmptyFormValue}
           rows={1}
           disabled={!canEdit}
         />
@@ -109,6 +115,7 @@ const ElectronicAccessFields = props => {
           aria-label={urlPublicNoteLabel}
           name={`${field}.publicNote`}
           component={TextArea}
+          parse={parseEmptyFormValue}
           rows={1}
           disabled={!canEdit}
         />